Update git workflow
