+++ .eslintignore
@@ -3,4 +3,4 @@ helix-importer-ui
 _src-lp/scripts/vendor/**/*.js
 sass-compile.js
 **/*.spec.js
-e2e
+ghosti
