+++ ghosti/json-tests/snapshot-block.js
